feat: Refactor view interfaces and implementations
The code changes include adding a new enum `ViewType` in the `ViewType.java` file, creating a new interface `PanelView` in the `PanelView.java` file, and deleting the `TableView.java` file. The `Main.java` file was modified to import the `ViewType` enum and update the method call to `displayView` in the `AdminOperationsController` class. The `CartView.java` file was deleted, and the `AppConfig.java` file was modified to update the `views` bean definition. The `AdminView.java`, `UserView.java`, and `ProductView.java` files were modified to implement the `PanelView` interface and update the method names.